Remove support for python versions < 3.6
