+++ gql/transport/local_schema.py
@@ -1,9 +1,6 @@
-from typing import Any, Union
-
 from graphql import GraphQLSchema
 from graphql.execution import ExecutionResult, execute
 from graphql.language.ast import Document
-from promise import Promise
 
 from gql.transport import Transport
 
@@ -12,8 +9,7 @@ class LocalSchemaTransport(Transport):
     """A transport for executing GraphQL queries against a local schema."""
 
     def __init__(
-        self,  # type: LocalSchemaTransport
-        schema,  # type: GraphQLSchema
+        self, schema: GraphQLSchema,
     ):
         """Initialize the transport with the given local schema.
 
@@ -21,8 +17,7 @@ class LocalSchemaTransport(Transport):
         """
         self.schema = schema
 
-    def execute(self, document, *args, **kwargs):
-        # type: (Document, *Any, **Any) -> Union[ExecutionResult, Promise[ExecutionResult]]
+    def execute(self, document: Document, *args, **kwargs) -> ExecutionResult:
         """Execute the given document against the configured local schema.
 
         :param document: GraphQL query as AST Node object.
@@ -30,4 +25,4 @@ class LocalSchemaTransport(Transport):
         :param kwargs: Keyword options passed to execute method from graphql-core library.
         :return: Either ExecutionResult or a Promise that resolves to ExecutionResult object.
         """
-        return execute(self.schema, document, *args, **kwargs)
+        return execute(self.schema, document, *args, **kwargs)  # type: ignore
