+++ gql/transport/requests.py
@@ -1,6 +1,4 @@
-from __future__ import absolute_import
-
-from typing import Any, Dict, Union
+from typing import Any, Dict, Optional, Union
 
 import requests
 from graphql.execution import ExecutionResult
@@ -20,17 +18,17 @@ class RequestsHTTPTransport(Transport):
     """
 
     def __init__(
-        self,  # type: RequestsHTTPTransport
-        url,  # type: str
-        headers=None,  # type: Dict[str, Any]
-        cookies=None,  # type: Union[Dict[str, Any], RequestsCookieJar]
-        auth=None,  # type: AuthBase
-        use_json=False,  # type: bool
-        timeout=None,  # type: int
-        verify=True,  # type: bool
-        retries=0,  # type: int
-        method="POST",  # type: str
-        **kwargs  # type: Any
+        self,
+        url: str,
+        headers: Optional[Dict[str, Any]] = None,
+        cookies: Optional[Union[Dict[str, Any], RequestsCookieJar]] = None,
+        auth: Optional[AuthBase] = None,
+        use_json: bool = False,
+        timeout: Optional[int] = None,
+        verify: bool = True,
+        retries: int = 0,
+        method: str = "POST",
+        **kwargs,
     ):
         """Initialize the transport with the given request parameters.
 
@@ -76,8 +74,12 @@ class RequestsHTTPTransport(Transport):
             for prefix in "http://", "https://":
                 self.session.mount(prefix, adapter)
 
-    def execute(self, document, variable_values=None, timeout=None):
-        # type: (Document, Dict, int) -> ExecutionResult
+    def execute(  # type: ignore
+        self,
+        document: Document,
+        variable_values: Optional[Dict[str, Any]] = None,
+        timeout: Optional[int] = None,
+    ) -> ExecutionResult:
         """Execute the provided document AST against the configured remote server.
         This uses the requests library to perform a HTTP POST request to the remote server.
