+++ gql/transport/websockets.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import
-
 import asyncio
 import json
 import logging
