+++ setup.cfg
@@ -1,15 +1,16 @@
 [wheel]
-universal=1
+universal = 1
 
 [flake8]
 max-line-length = 120
 
 [isort]
-known_standard_library=ssl
-known_first_party=gql
-multi_line_output=3
-include_trailing_comma=True
-line_length=88
+known_standard_library = ssl
+known_first_party = gql
+multi_line_output = 3
+include_trailing_comma = True
+line_length = 88
+not_skip = __init__.py
 
 [tool:pytest]
 norecursedirs = venv .venv .tox .git .cache .mypy_cache .pytest_cache
