+++ setup.py
@@ -1,48 +1,34 @@
-import sys
-
 from setuptools import setup, find_packages
 
 install_requires = [
-    "six>=1.10.0",
+    "aiohttp==3.6.2",
     "graphql-core>=2.3.2,<3",
-    "promise>=2.3,<3",
     "requests>=2.12,<3",
+    "six>=1.10.0",
+    "websockets>=8.1,<9",
+    "yarl>=1.0,<2.0",
 ]
 
-scripts = []
+scripts = [
+    "scripts/gql-cli",
+]
 
-if sys.version_info > (3, 6):
-    install_requires.append(
-        ["websockets>=8.1,<9", "aiohttp==3.6.2", "yarl>=1.0,<2.0",]
-    )
-    scripts.append("scripts/gql-cli")
-
-tests_require = (
-    [
-        "pytest==5.4.2",
-        "pytest-asyncio==0.11.0",
-        "pytest-cov==2.8.1",
-        "mock==4.0.2",
-        "vcrpy==4.0.2",
-        "coveralls==2.0.0",
-        "parse>=1.6.0",
-    ]
-    if sys.version_info > (3, 6)
-    else [
-        "pytest==4.6.9",
-        "pytest-cov==2.8.1",
-        "vcrpy==3.0.0",
-        "mock==3.0.0",
-        "coveralls==1.11.1",
-    ]
-)
+tests_require = [
+    "coveralls==2.0.0",
+    "parse>=1.6.0",
+    "pytest==5.4.2",
+    "pytest-asyncio==0.11.0",
+    "pytest-cov==2.8.1",
+    "mock==4.0.2",
+    "vcrpy==4.0.2",
+]
 
 dev_requires = [
+    "black==19.10b0",
+    "check-manifest>=0.40,<1",
     "flake8==3.7.9",
     "isort==4.2.8",
-    "black==19.10b0",
     "mypy==0.770",
-    "check-manifest>=0.40,<1",
 ] + tests_require
 
 setup(
