+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist =
     black,flake8,import-order,mypy,manifest,
-    py{27,35,36,37,38,39-dev,py,py3},cov
+    py{36,37,38,39-dev,py3}
 ; requires = tox-conda
 
 [pytest]
@@ -13,7 +13,6 @@ setenv =
     PYTHONPATH = {toxinidir}
     MULTIDICT_NO_EXTENSIONS = 1 ; Related to https://github.com/aio-libs/multidict
     YARL_NO_EXTENSIONS = 1      ; Related to https://github.com/aio-libs/yarl
-    COVERAGE_FILE=.coverage.{envname}
 install_command = python -m pip install --ignore-installed {opts} {packages}
 whitelist_externals =
     python
@@ -21,12 +20,8 @@ deps = -e.[test]
 ; Prevent installing issues: https://github.com/ContinuumIO/anaconda-issues/issues/542
 commands =
     pip install -U setuptools
-    py{27}: pytest {posargs:tests --cov-report=term-missing --cov=gql -s}
-    py{35,py}: pytest {posargs:tests -s}
     py{36,37,39-dev,py3}: pytest {posargs:tests tests_py36 -s}
     py{38}: pytest {posargs:tests tests_py36 --cov-report=term-missing --cov=gql -s}
-    cov: coverage combine {toxinidir}/.coverage.py27 {toxinidir}/.coverage.py38
-    cov: coverage report --show-missing
 
 [testenv:black]
 basepython=python3.8
